consider ASSET_DIVIDEND transactions when computing return
When computing the returns of an asset, transactions of type
ASSET_DIVIDEND are not considered. That makes the return rate computed
by beangrow substantially different to rates reported by fund managers.

I could not find any justification as to why these transactions are
ignored when computing the return rates, and this seems to confuse me as
well as other users. See beancount/beangrow#23 and beancount/beangrow#26

At least with this small patch, users get a chance to try and see if
they get the result they expected.